+++ .gitignore
@@ -1,0 +1,8 @@
+﻿################################################################################
+# 此 .gitignore 文件已由 Microsoft(R) Visual Studio 自动创建。
+################################################################################
+
+/lib/net48
+/.vs
+/AudioCore/obj
+/AudioApi/obj
