--- AudioCore/EventArgs/Voice/TrackSelectedEventArgs.cs
@@ -1,17 +1,0 @@
-﻿namespace AudioApi.AudioCore.EventArgs.Voice
-{    /// <summary>
-     /// 音频选择完成
-     /// </summary>
-    public class TrackSelectedEventArgs
-    {
-        public VoicePlayerBase VoicePlayerBase { get; }
-        public bool DirectPlay { get; }
-        public int QueuePos { get; }
-        public TrackSelectedEventArgs(VoicePlayerBase playerBase, bool directPlay, int queuePos)
-        {
-            VoicePlayerBase = playerBase;
-            DirectPlay = directPlay;
-            QueuePos = queuePos;
-        }
-    }
-}
